+++ Sources/Sentry/SentryTimeToDisplayTracker.m
@@ -86,9 +86,9 @@
     [tracer setShouldIgnoreWaitForChildrenCallback:^(id<SentrySpan> span) {
         if ([span.origin isEqualToString:SentryTraceOriginAutoUITimeToDisplay]) {
             return YES;
-        } else {
-            return NO;
         }
+
+        return NO;
     }];
     [tracer setFinishCallback:^(SentryTracer *_tracer) {
         [SentryDependencyContainer.sharedInstance.framesTracker removeListener:self];
@@ -143,12 +143,23 @@
     }
 
     if (self.fullDisplaySpan.isFinished == NO) {
+        if (_fullyDisplayedReported) {
+            SENTRY_LOG_DEBUG(
+                @"SentrySDK.reportFullyDisplayed() was called but didn't receive a new frame to "
+                @"finish the TTFD span. Finishing the full display span so the SDK can start a new "
+                @"time to display tracker.");
+            [self.fullDisplaySpan finish];
+            return;
+        }
+
         SENTRY_LOG_WARN(@"You didn't call SentrySDK.reportFullyDisplayed() for UIViewController: "
                         @"%@. Finishing full display span with status: %@.",
             _name, nameForSentrySpanStatus(kSentrySpanStatusDeadlineExceeded));
 
         [self.fullDisplaySpan finishWithStatus:kSentrySpanStatusDeadlineExceeded];
     }
+
+    [SentryDependencyContainer.sharedInstance.framesTracker removeListener:self];
 }
 
 - (void)framesTrackerHasNewFrame:(NSDate *)newFrameDate